auth

+++ 1-glossary/server/wordRoutes.js
@@ -3,12 +3,14 @@ const router = require('express').Router();
 module.exports.wordRoutes = function (db) {
   router.post('/', async function (req, res) {
     const { word, definition } = req.body;
+    const userId = req.userId;
+
     if (!word || !definition)
       return res.sendStatus('500');
 
     try {
       const newWord = await new db.Words({
-        word, definition
+        word, definition, userId
       }).save();
       res.status(201).send();
     } catch (err) {
@@ -18,11 +20,12 @@ module.exports.wordRoutes = function (db) {
   });
 
   router.get('/', async function (req, res) {
-    let query = {};
+    const userId = req.userId;
+    let query = { userId };
     let page = Number(req.query.skip) || 0;
 
     if (req.query.q)
-      query = { word: new RegExp(req.query.q, 'i') };
+      query['word'] = new RegExp(req.query.q, 'i');
 
     const data = await db.Words
       .find(query)
@@ -31,7 +34,9 @@ module.exports.wordRoutes = function (db) {
       .limit(10)
       .exec();
 
-    const total = await db.Words.count();
+    const total = await db.Words
+      .find({ userId })
+      .count();
 
     res.send({ page, total, data })
   });
@@ -51,7 +56,7 @@ module.exports.wordRoutes = function (db) {
 
   router.delete('/:id', async function (req, res) {
     try {
-      const result = await db.Words.deleteOne({ _id: req.params.id });
+      const result = await db.Words.deleteOne({ _id: req.params.id, userId: req.userId });
       res.send(result);
     } catch (err) {
       console.error(err);
